initial upload of apk
to install, go to android settings ---> security ---> google play protect, and turn the scanning to off. then go to the .apk file in downloads folder, tap it, and install the app. it will not automatically creates a shortcut, so you must add it manually, or just go to apps and you will see it there. this was developed for, and tested on android 9.0. it should work for kindle fire 5th gen and up, and other androids.